test(adapter-api): move PR-detection parser tests
Move the parser's unit-test module out of mainframe-adapter-claude
into an integration test against the new adapter-neutral module
(todo #339 task 2); assertion-for-assertion, 28 tests pass.
